feat: add header filters to remaining columns in Incoming, Preparation, Load FCs, Pooling, and Stats
Extracted invoicingConsts.js's numeric comparison-expression filter (>100, 50-200, etc.) into a shared frontend/src/utilities/numericHeaderFilter.js so every module's numeric/QC/cost columns get the same syntax instead of duplicating it per file.

Plain-text columns (Protocol, Comment, Pool, Index IDs, Coord, etc.) get a substring headerFilter with a gray "Filter..." hint. The two categorical columns that were free text or explicitly disabled (Unit facility, GMO facility in Incoming/Preparation) get a headerFilter: "select" dropdown built from their existing editor value lists instead.
