+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 
 - Libraries and Samples: every column search box (Status, Name, Type, Plate Coord, Barcode, Pool Paths, GMO, Date, Input Type, Comment, Organism, Protocol, Analysis Type, Coord, Index I7/I5, Length, Flowcell IDs, Sequencers, plus the existing Index Type/I7 ID/I5 ID) now filters server-side across the full result set, not just the loaded page, with gray hint text/tooltips showing the expected syntax. Plate Coord filters against the server-computed value from #339, matched by re-deriving it for every request name still passing the other active filters. The Status/Protocol/Analysis Type/Sequencer/Read Length dropdowns are back in Advanced Filters too, now as shortcuts onto the same column search boxes (picking a dropdown value fills in and stays in sync with the matching column's box) rather than a separate ID-based filter path. The server-side filter debounce was bumped from 500ms to 800ms to avoid reloading mid-typing.
 - Column search boxes across Incoming, Preparation, Pooling, Load FCs, and the Stats tables now show gray "Filter..." hint text, matching the pattern already used in Invoicing.
+- Added header-filter search boxes to nearly every remaining column across Incoming, Preparation, Load FCs, Pooling, Run Statistics, and Sequences Statistics: numeric/QC/cost columns get Invoicing's comparison syntax (`>100`, `50-200`, etc. — now shared via `numericHeaderFilter.js` instead of duplicated per module), plain-text columns (Protocol, Comment, Index IDs, Coord, etc.) get a substring filter, and the two categorical columns (Unit facility, GMO facility) get a dropdown (`headerFilter: "select"`) instead of free text.
 - Libraries & Samples: "Plate Coord" (A1..H12) now computes on the server (`CompleteLibraryData`/`CompleteSampleData.plate_coord`, served by `/api/libraries_and_samples/`), not in the browser. It ranks libraries and samples together by barcode within each request name. A check against every real request in a production snapshot found one bug. A barcode-only key mishandled two 2019 Sample records with the same legacy barcode. The fix instead keys on request name, record type, and primary key.
 - Added I7 Index / I5 Index range filters (From/To) to Libraries & Samples Advanced Filters, so users can find libraries/samples by index ID range (e.g. N701 to N729) without relying on exact Index Type name matches.
 - Fixed Libraries & Samples search returning no results for an Index Type that exists in the data: the global search endpoint never queried `index_type_name`.
