+++ frontend/src/constants/loadFlowcellsConsts.js
@@ -7,6 +7,7 @@ import iconSelectAll from "../assets/icons/action_select_all.svg";
 import iconDeselectAll from "../assets/icons/action_deselect_all.svg";
 import iconDestroyPool from "../assets/icons/action_pool_destroy.svg";
 import iconExportDownload from "../assets/icons/export_download.svg";
+import { numericFilterConfig } from "../utilities/numericHeaderFilter";
 
 export function loadFlowcellsGroupHeader(value, rows = []) {
   const formattedDate = rows[0]?.create_time || "";
@@ -179,6 +180,7 @@ export function loadFlowcellsColumnDefs(getTabulatorInstance, callbacks = {}) {
       minWidth: 110,
       width: 120,
       visible: true,
+      ...numericFilterConfig((v) => Number(v)),
       editor: "number",
       editorParams: {
         min: 0,
@@ -204,6 +206,7 @@ export function loadFlowcellsColumnDefs(getTabulatorInstance, callbacks = {}) {
       minWidth: 90,
       width: 95,
       visible: true,
+      ...numericFilterConfig((v) => Number(v)),
       editor: "number",
       editorParams: {
         min: 0,
